+++ src/main/java/ir/maktabsharif/finalproject/service/impl/SpecialistServiceImpl.java
@@ -1,24 +1,19 @@
 package ir.maktabsharif.finalproject.service.impl;
 
 
-import ir.maktabsharif.finalproject.entities.Order;
-import ir.maktabsharif.finalproject.entities.Specialist;
-import ir.maktabsharif.finalproject.entities.SubTask;
-import ir.maktabsharif.finalproject.entities.Suggestions;
-import ir.maktabsharif.finalproject.enumerations.OrderStatus;
+import ir.maktabsharif.finalproject.dto.SpecialistDto;
+import ir.maktabsharif.finalproject.entities.*;
+import ir.maktabsharif.finalproject.enumerations.Role;
 import ir.maktabsharif.finalproject.enumerations.SpecialistStatus;
+import ir.maktabsharif.finalproject.exception.SpecialistNotFoundException;
 import ir.maktabsharif.finalproject.exception.SpecialistOperationException;
 import ir.maktabsharif.finalproject.repository.SpecialistRepository;
-import ir.maktabsharif.finalproject.service.OrderService;
-import ir.maktabsharif.finalproject.service.SpecialistService;
-import ir.maktabsharif.finalproject.service.SuggestionsService;
+import ir.maktabsharif.finalproject.service.*;
+import ir.maktabsharif.finalproject.util.MapperUtil;
 import ir.maktabsharif.finalproject.util.ValidationUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 
 @Service
@@ -26,8 +21,26 @@ import java.util.List;
 public class SpecialistServiceImpl implements SpecialistService {
     private final ValidationUtil validationUtil;
     private final SpecialistRepository specialistRepository;
-    private SuggestionsService suggestionsService;
-    private OrderService orderService;
+    private final SubTaskService subTaskService;
+    private final UsersService usersService;
+    private final MapperUtil mapperUtil;
+
+    @Override
+    public SpecialistDto signUp(String firstName, String lastName, String email, String username, String password) throws SpecialistOperationException {
+            Specialist specialist = Specialist.builder()
+                    .firstName(firstName)
+                    .lastName(lastName)
+                    .email(email)
+                    .username(username)
+                    .password(password)
+                    .role(Role.SPECIALIST)
+                    .specialistStatus(SpecialistStatus.PENDING_APPROVAL)
+                    .score(0.0)
+                    .build();
+            add(specialist);
+            return mapperUtil.convertToDto(specialist);
+    }
+
     @Override
     public void add(Specialist specialist) throws SpecialistOperationException {
         try {
@@ -39,7 +52,6 @@ public class SpecialistServiceImpl implements SpecialistService {
         } catch (Exception e) {
             throw new SpecialistOperationException("An error occured while adding specialist", e);
         }
-
     }
 
     @Override
@@ -53,9 +65,14 @@ public class SpecialistServiceImpl implements SpecialistService {
     }
 
     @Override
-    public void delete(Specialist specialist) throws SpecialistOperationException {
+    public void delete(SpecialistDto specialistDto) throws SpecialistOperationException {
         try {
-            specialistRepository.delete(specialist);
+            if (doesSpecialistExist(specialistDto)) {
+                Specialist specialist = specialistRepository.findByFirstNameAndLastName(specialistDto.firstname(), specialistDto.lastname());
+                specialistRepository.delete(specialist);
+            } else {
+                throw new SpecialistNotFoundException("Specialist not found");
+            }
         } catch (Exception e) {
             throw new SpecialistOperationException("An error occured while deleting specialist", e);
         }
@@ -89,48 +106,26 @@ public class SpecialistServiceImpl implements SpecialistService {
     }
 
     @Override
-    public boolean checkSpecialistImage(Specialist specialist) throws SpecialistOperationException {
+    public boolean checkSpecialistImage(SpecialistDto specialistDto) throws SpecialistOperationException {
         try {
-            return specialist.getPersonalImage() != null;
+            Specialist specialist =
+                    specialistRepository.findByFirstNameAndLastName(specialistDto.firstname(), specialistDto.lastname());
+            if (doesSpecialistExist(specialistDto)) {
+                return specialist.getPersonalImage() != null;
+            } else {
+                throw new SpecialistNotFoundException("Specialist not found");
+            }
         } catch (Exception e) {
             throw new SpecialistOperationException("An error occured while checking specialist personal image", e);
         }
     }
 
+
     @Override
-    public Suggestions sendSuggestionsForRelatedSubTasks(Specialist specialist, Order order, Double suggestedPrice, LocalDate dateOfService, LocalTime timeOfWork) throws SpecialistOperationException {
-        try {
-            List<SubTask> specialistSubTask = specialist.getSubTasks();
-            if (canRespondToSuggestion(specialistSubTask,specialist,order)) {
-                if(suggestedPrice >= order.getSubTask().getBasePrice() && dateOfService.isAfter(order.getCreationDate())){
-                    suggestionsService.add(Suggestions.builder()
-                            .specialist(specialist)
-                            .suggestedPrice(suggestedPrice)
-                            .workTime(timeOfWork)
-                            .suggestedDate(dateOfService)
-                            .order(order)
-                            .build());
-                    order.setStatus(OrderStatus.WAITING_FOR_SPECIALIST_SELECTION);
-                    orderService.update(order);
-                }else{
-                    System.out.println("You dont have any order to send suggestion ! ");
-                    return null;
-                }
-            }else{
-                System.out.println("You dont have any order to send suggestion ! ");
-                return null;
-            }
-        } catch (Exception e) {
-            throw new SpecialistOperationException("An error occured while sending suggestion from specialist", e);
-        }
-        return null;
+    public boolean doesSpecialistExist(SpecialistDto specialistDto) throws SpecialistOperationException {
+        Specialist specialist =
+                findByFirstNameAndLastName(mapperUtil.convertToEntity(specialistDto).getFirstName(), mapperUtil.convertToEntity(specialistDto).getLastName());
+        return specialist.getId() != null && specialist.getFirstName() != null && specialist.getLastName() != null;
     }
 
-    private boolean canRespondToSuggestion(List<SubTask> specialistSubTask, Specialist specialist, Order order) {
-        return !specialistSubTask.isEmpty() &&
-                !(specialist.getSpecialistStatus().equals(SpecialistStatus.PENDING_APPROVAL)
-                        || specialist.getSpecialistStatus().equals(SpecialistStatus.NOT_APPROVED))
-                && order.getStatus().equals(OrderStatus.WAITING_FOR_SPECIALIST_SELECTION) ||
-                order.getStatus().equals(OrderStatus.WAITING_FOR_SPECIALIST_PROPOSALS);
-    }
 }
